Rename entity
